# Port `perf.ts`