Port `perf.ts`. Ran the test on my machine:

### v1:
```
large google.protobuf.FileDescriptorSet (1113407 bytes) x 60.06 ops/sec ±0.89% (63 runs sampled)
tiny docs.User (4 bytes) x 3,422,754 ops/sec ±0.58% (95 runs sampled)
scalar values (102 bytes) x 757,654 ops/sec ±0.28% (94 runs sampled)
repeated scalar fields (195 bytes) x 460,627 ops/sec ±0.24% (98 runs sampled)
map with scalar keys and values (186 bytes) x 239,602 ops/sec ±0.25% (96 runs sampled)
repeated field with 1000 messages (2000 bytes) x 9,223 ops/sec ±0.25% (99 runs sampled)
map field with 1000 messages (8890 bytes) x 3,677 ops/sec ±0.47% (94 runs sampled)
```
### v2:
```
large google.protobuf.FileDescriptorSet (1113407 bytes) x 6.67 ops/sec ±1.61% (21 runs sampled)
tiny docs.User (4 bytes) x 659,333 ops/sec ±0.71% (94 runs sampled)
scalar values (102 bytes) x 77,491 ops/sec ±0.76% (94 runs sampled)
repeated scalar fields (195 bytes) x 46,473 ops/sec ±0.70% (95 runs sampled)
map with scalar keys and values (186 bytes) x 48,238 ops/sec ±0.73% (98 runs sampled)
repeated field with 1000 messages (2000 bytes) x 576 ops/sec ±0.48% (93 runs sampled)
map field with 1000 messages (8890 bytes) x 144 ops/sec ±0.43% (83 runs sampled)
```